requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
python-dotenv==0.19.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
alembic==1.7.1
python-dateutil==2.8.2 